actorFaceRecognition/requirements.txt
gradio==3.39.0
opencv-python
seaborn
Pillow
scikit-learn==1.3.0
numpy
PyWavelets
joblib